Fixed ultimate UI bug and moved test scenes to test folder

+++ Assets/Scripts/Managers/GameUIManager.cs
@@ -147,9 +147,8 @@ class GameUIManager : Script
         {
             return;
         }
-
         Vector2 textureCoordinates = ultimateBar.textureCoordinatesRange;
-        textureCoordinates.x = Mathf.Interpolate(0, 1f, currentSp / maxSp, 1);
+        textureCoordinates.x = Mathf.Interpolate(0, 1f, (float)currentSp / maxSp, 1);
         ultimateBar.textureCoordinatesRange = textureCoordinates;
     }
